e2e test transparent SA rewrite in downstream deployment
deploy a demo application via KCP to a downstream cluster. this application explicitely uses the in-cluster configuration to create a configmap in an upstream KCP namespace. the e2e test verifies, that this configmap gets created.
this part of the e2e test only runs when a workload cluster is available, e.g. during CI or when the testflags `--pcluster-kubeconfig`and `--icc-test-image` are present. the GH-action CI provides a freshly build in-cluster test image automatically. a prebuild multi-arch image is available under ghcr.io/kcp-dev/kcp/icc-test-app:main

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>